# api/data_explorer/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "API Service that reads from Elasticsearch."
  version: "0.0.1"
  title: "Data Explorer Service"
  license:
    name: "BSD"
    url: "http://opensource.org/licenses/BSD-3-Clause"
paths:
...
  /exportUrl:
    post:
      tags:
      - "ExportUrl"
      description: "Creates and returns a signed URL to a GCS zip file of JSON files.\
        \ The JSON files represent entities to be exported to a Saturn workspace.\
        \ https://bvdp-saturn-prod.appspot.com/#import-data may be called with the\
        \ url parameter set to this url. For each JSON file, https://rawls.dsde-prod.broadinstitute.org/#!/entities/create_entity\
        \ may be called with the JSON as the POST body.\n"
      operationId: "export_url_post"
      parameters:
      - in: "body"
        name: "exportUrlRequest"
        required: false
        schema:
          $ref: "#/definitions/exportUrlRequest"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/ExportUrlResponse"
      x-swagger-router-controller: "data_explorer.controllers.export_url_controller"
definitions:
...
  ExportUrlResponse:
    properties:
      url:
        type: "string"
        description: "URL-encoded signed URL of GCS file containing entities to export."
      authorization_domain:
        type: "string"
        description: "Optional FireCloud authorization domain"
    description: "Information for sending data to Saturn"
    example:
      authorization_domain: "authorization_domain"
      url: "url"
...
  exportUrlRequest:
    type: "object"
    properties:
      cohortName:
        type: "string"
      filter:
        type: "array"
        items:
          type: "string"